refactor(components): share rule table controls and row ops across pages
RuleTableControls and makeRowOps were living inside pages/JiraTemplates but
are needed verbatim by the Guardrails rules table. Graduate both to the
shared layer instead of duplicating them:

- pages/JiraTemplates/Form/components/RuleTableControls.jsx -> src/components/RuleTableControls/
- makeRowOps -> src/utils/rowOps.js

No behavior change; the four Jira Templates tables keep the same props.
