+++ webapp_v2/src/pages/JiraTemplates/helpers.js
@@ -145,42 +145,6 @@ export function apiTemplateToCmdbRows(template) {
   return rows.length ? rows : [createEmptyCmdbRow()]
 }
 
-// Row operations shared by the four rule tables. The two mapping tables pass a
-// filterFn because they render disjoint subsets of one shared rows array —
-// select/delete must only touch the rows the table actually shows.
-export function makeRowOps({ rows, setRows, factory, filterFn = () => true }) {
-  const visible = rows.filter(filterFn)
-  const allSelected = visible.length > 0 && visible.every((r) => r.selected)
-  return {
-    visible,
-    allSelected,
-    patchRow: (id, patch) =>
-      setRows((current) =>
-        current.map((r) => (r.id === id ? { ...r, ...patch } : r)),
-      ),
-    toggleSelect: (id) =>
-      setRows((current) =>
-        current.map((r) => (r.id === id ? { ...r, selected: !r.selected } : r)),
-      ),
-    toggleAll: () =>
-      setRows((current) =>
-        current.map((r) =>
-          filterFn(r) ? { ...r, selected: !allSelected } : r,
-        ),
-      ),
-    deleteSelected: () =>
-      setRows((current) => {
-        const remaining = current.filter((r) => !(filterFn(r) && r.selected))
-        return remaining.length ? remaining : [factory()]
-      }),
-    addRow: (transform) =>
-      setRows((current) => [
-        ...current,
-        transform ? transform(factory()) : factory(),
-      ]),
-  }
-}
-
 // Incomplete rows are dropped on submit (same criteria as the legacy app);
 // UI-only keys (id, selected) never reach the payload.
 function prepareMappingItems(rows) {
